Change 6N3P to 6N17B
Because the correction circuit uses a differential amplifier, its gain
is devided by 2. With µ=35 it makes a 6N3P to provide a gain between 10
and 15. In order to get a good enough damping factor, a higher gain is
desirable while keeping the CMRR and PSRR of the 6N3P. The 6N17B has
been found as beeing a suitable replacement giving an estimated gain of
28 with a comparable ripple rejection ratio and with a better protection
against microphony.