Replace CDiff with 6N3P µ-follower
As stated, the differential pair was not worth two tubes to present
normal PSRR and CMRR to the 0V. A FET driven µ-follower has been
prefered. It presents a load of 500k and an output impedance of less
than 100Ω while keeping noise away from the FET’s source. The µ-follower
has also the advantage of having two outputs: one (used here) is the
FET’s source and the other is the high impedance valve’s anode. This is
a perfect place to set a damping circuit if needed.