Rearrange PSU and drivers
With 300Vac output at main transformer, we are able to get a regulated
330Vdc for drivers. Still, it presents a 10mV ripple @ 100Hz. The
regulated 200Vdc does not present such ripple. The CDiff can be powered
from the 200V since it uses a FET µ-follower.

PSU currents have been simulated to check the amplifier can stand
maximum output without the PSU to choke.

The regulation is now preceded by a RC filter that lowers the ripple of
6dB.